Implement AppHeader and AppNavigation components; refactor Layout to utilize them for improved structure and maintainability
